refactor(katana): rename module
commit-id:5c8b2502
