+++ crates/katana/core/src/backend/block.rs
