+++ crates/katana/core/src/backend/contract.rs
