+++ crates/katana/core/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod accounts;
+pub mod backend;
 pub mod block_context;
 pub mod constants;
 pub mod db;
 pub mod sequencer;
-pub mod starknet;
 pub mod state;
 pub mod util;
