+++ crates/katana/core/src/sequencer.rs
@@ -24,13 +24,13 @@ use starknet_api::transaction::{
 use tokio::sync::{RwLock, RwLockReadGuard, RwLockWriteGuard};
 use tokio::time;
 
+use crate::backend::block::StarknetBlock;
+use crate::backend::config::StarknetConfig;
+use crate::backend::contract::StarknetContract;
+use crate::backend::event::EmittedEvent;
+use crate::backend::transaction::ExternalFunctionCall;
+use crate::backend::StarknetWrapper;
 use crate::sequencer_error::SequencerError;
-use crate::starknet::block::StarknetBlock;
-use crate::starknet::config::StarknetConfig;
-use crate::starknet::contract::StarknetContract;
-use crate::starknet::event::EmittedEvent;
-use crate::starknet::transaction::ExternalFunctionCall;
-use crate::starknet::StarknetWrapper;
 use crate::state::DictStateReader;
 use crate::util::starkfelt_to_u128;
